Fixed issue: [security] Possible SQL injection in data entry, CPDB participant sharing and condition editing - found by James Mullen, Edgescan

+++ application/controllers/admin/participantsaction.php
@@ -2127,7 +2127,7 @@ $url .= "_view"; });
           'isSuperAdmin' => $isSuperAdmin
         ];
         $participantIds = Yii::app()->request->getPost('participant_id');
-        $iShareUserId = Yii::app()->request->getPost('shareuser');
+        $iShareUserId = (int) Yii::app()->request->getPost('shareuser');
         $bCanEdit = Yii::app()->request->getPost('can_edit') == 'on';
 
         if (!is_array($participantIds)) {
@@ -2145,7 +2145,7 @@ $url .= "_view"; });
             foreach ($participantIds as $id) {
                 $time = time();
                 $aData = array(
-                    'participant_id' => $id,
+                    'participant_id' => (int) $id,
                     'share_uid' => $iShareUserId,
                     'date_added' => date('Y-m-d H:i:s', $time),
                     'can_edit' => $bCanEdit
